SIHFinals/SIHProject/requirements.txt
ollama
numpy
pandas
tqdm
flask
flask-cors
python-dotenv
gunicorn
